% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allplot.R
\name{draw_all_plot}
\alias{draw_all_plot}
\title{Descriptive Plots}
\usage{
draw_all_plot(your_variable, your_dataset)
}
\arguments{
\item{your_variable}{This should be the name of the variable you want to plot}

\item{your_dataset}{This should be the dataset containing the target variable}
}
\value{
Returns the histogram, quantile plots and boxplots of your variable
}
\description{
Descriptive Plots
}
