% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fred.R
\name{fred_get_series}
\alias{fred_get_series}
\title{Get Fred Data}
\usage{
fred_get_series(APIkey, search_text, realtime_start)
}
\arguments{
\item{APIkey}{This should be the API key of the Fred page}

\item{search_text}{This should be the key word of the data you want to get. The default is "interest+rate"}

\item{realtime_start}{This should be the starting data, the format should be yyyy-mm-dd and the default is "2000-01-01"}
}
\value{
Returns the data frame with the data you want
}
\description{
Get Fred Data
}
\examples{
fred_get_series(your_API_Key, search_text="interest+rate", realtime_start = "2000-01-01")
}
