% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometric_mean.R
\name{geom_return}
\alias{geom_return}
\title{Geometric Mean}
\usage{
geom_return(your_return)
}
\arguments{
\item{your_return}{this should be a data frame consisting of return data}
}
\value{
Returns the geometric mean of the data input.
}
\description{
Geometric Mean
}
