% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_impute.R
\name{knn_impute}
\alias{knn_impute}
\title{KNN Imputation}
\usage{
knn_impute(target_variable, your_dataset)
}
\arguments{
\item{target_variable}{This should be the name of the variable you want to impute}

\item{your_dataset}{This should be the dataset containing the target variable}
}
\value{
Returns the whole data frame with the missing value of the target variable being imputed
}
\description{
KNN Imputation
}
