% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_window_plot.R
\name{rolling_plot}
\alias{rolling_plot}
\title{Rolling Window Plot}
\usage{
rolling_plot(your_x_variable, your_y_variable, your_dataset)
}
\arguments{
\item{your_x_variable}{this should be a data frame}

\item{your_y_variable}{this should be a data frame}

\item{your_dataset}{this should be a data frame containing all the variables you want to analyse}
}
\value{
Return the rolling window plot of the relationship between x and y
}
\description{
Rolling Window Plot
}
